Repair build purity CI contracts
Keep semantic library coverage on the explicit compatibility backend, preserve Cranelift-specific runtime tests, teach proof and property gates to accept only stable fail-closed lowering, and extract helpers to satisfy compiler decomposition ratchets.
